Command infrastructure for propulsion module
Setting up the command infrastructure so that propulsion can be
commanded.  The command payload will use 8 bytes to representing le
floats. Each float is a vector component.

The vector will be used to determine the pwm duty cycle for each of the
three thrusters in order to achieve thrust in the direction of the
vector, where y is parallel to the vehicle, and x is perpendicular.